+++ R/0_ProposeTessellation.R
@@ -11,8 +11,8 @@
 #'   centre and each column corresponds to a dimension.
 #' @param dim_j An integer vector specifying the column indices of the covariates
 #'   that define the dimensions of the current tessellation.
-#' @param var The variance parameter for the normal distribution (`rnorm`) used to
-#'   generate new coordinate values for centres and dimensions.
+#' @param sd The standard deviation parameter for the normal distribution (`rnorm`) 
+#'   used to generate new coordinate values for centres and dimensions.
 #' @param covariateIndices This parameter is accepted by the function but is
 #'   **not used** in the current implementation.
 #' @param NumCovariates An integer giving the total number of covariates
@@ -29,13 +29,13 @@
 #' @noRd
 #' 
 #' @useDynLib AddiVortes, .registration = TRUE
-proposeTessellation <- function(tess_j, dim_j, var, covariateIndices,
+proposeTessellation <- function(tess_j, dim_j, sd, covariateIndices,
                                 NumCovariates) {
   # Call the C++ implementation via the .Call interface
   results <- .Call("propose_tessellation_cpp",
                    tess_j,
                    dim_j,
-                   as.double(var),
+                   as.double(sd),
                    as.integer(NumCovariates))
   
   return(results)